Add workflow dispatch and replace project reference with NuGet package
Added workflow_dispatch to dotnet.yml to allow manual triggering of GitHub Actions. Also, replaced the Stxima.SendPulseClient project reference with a NuGet package reference in various .csproj files. This was done because the project is now available as a NuGet package, which improves the modularity and independence of our project structure. Additionally, removed reference to the aforementioned project from the solution file.