Switched to SQL database and reorganized data-related files
This commit changes the database usage in the code from PostgreSQL to SQL, due to a decision to switch database vendors. All instances where the PostgreSQL database was previously referenced in the code have been replaced with SQL database references. Also, data-related files such as Migrations and Configurations were moved into a new Sql directory to reflect this change. The ItemType class was moved to the Enums directory in the Domain layer for better organization. 'jsonb' type in the ItemEntityTypeConfiguration and ClientItemTypeConfiguration classes was replaced with 'text' for readability. ApplicationConfiguration now includes a DatabaseType to specify the type of database being used. Too long references in EntitiesProcessor were shortened for readability.