Add triggers on master branch push in workflow file
The changes in the master_azfun-stushbr-payment.yml file add a trigger for the build and deployment process.
This trigger is activated whenever there is a push to the master branch. The reason for the addition is
to ensure that any changes pushed to the master branch are automatically built and deployed, enhancing
the CI/CD process in the project.