Fix/function (#15)
* Add Stushbr.Api dependencies in Payment service

Added Stushbr.Api extensions in the Startup.cs file of Payment service and updated project references in the .csproj file. This was necessary to leverage utilities provided by the Stushbr.Api in the Payment service. This change leads to a cleaner and more organized code structure in the project.

* Update deployment workflow for Azure Function

The azure function application deployment workflow has been updated. The changes include the specification of the exact path to the project within the monorepo. This will prevent the pipeline from using the repository root as the default location. A caching layer has been added to save Nuget packages for more efficient builds. In addition, a new Nuget source has been configured to provide access to necessary private packages. The slot name for deployment has been also corrected to lower case to align with Azure naming conventions.

* Add triggers on master branch push in workflow file

The changes in the master_azfun-stushbr-payment.yml file add a trigger for the build and deployment process.
This trigger is activated whenever there is a push to the master branch. The reason for the addition is
to ensure that any changes pushed to the master branch are automatically built and deployed, enhancing
the CI/CD process in the project.

* Optimized caching for .NET dependencies in GitHub Actions

The GitHub Actions workflows for .NET have been modified to improve caching of dependencies, which can potentially speed up build times. Instead of caching only the NuGet packages, the workflows now cache both NuGet packages and other .NET dependencies. The caching step has been generalized and now uses 'actions/cache@v2' rather than 'actions/cache@v3'. The files which changes to should invalidate the cache has also been expanded to include *.csproj files, in addition to the existing packages.lock.json. This change should make the cache more precise, minimizing unnecessary cache invalidation.

* Update caching in GitHub workflows

Updated the 'Cache dependencies' step in two GitHub workflow files. The action `actions/cache@v2` was replaced with `actions/cache@v3` for better cache efficiency and performance. In addition, the path and key for caching were simplified to only target '.nuget/packages', ensuring only critical dependencies are being cached. The 'dotnet' caching that was previously included seemed unnecessary and was therefore excluded, reducing complexity in our workflow files.

* Update deployment name in workflow file

The deployment name for the Azure function app in the GitHub workflow file (.github/workflows/master_azfun-stushbr-payment.yml) was updated. This was done to make the name simpler and more descriptive, improving clarity and readability.