Removed unused modules and restructured project
Removed multiple unused modules that were previously part of the project. This change was made to declutter the codebase, improve clarity and simplifies future development. Additionally, some renaming and restructuring of the classes and files were performed to make the code more consistent and easier to navigate. A few new classes for handling queries were also added as part of this commit.