Refactor payment processing and add database migration features (#21)
* Refactor payment processing and add database migration features

Major refactor of the payment processing system to support the addition of 'TildaPayments'. This change included renaming and expanding existing classes, as well as creating new ones, to better handle different types of transactions. A new configuration for 'Tilda' was also added for enhanced integration.

Additionally, changes were made to the database structure, including updating the 'Item' entity, and adding a mechanism to handle automatic database migrations. This will ensure our database structure stays consistent and up-to-date with our application's needs. All changes aim to improve extensibility and maintainability of our codebase.

* remove automatic deployment trigger for azure function