Extract item card HTML into a partial view
This commit introduces a new partial view for the item card in the Items list. This has been done to make the code more modular and easier to maintain. Additionally, the HTML code for the modal in "TelegramItemDetails" was moved from the ItemsList view to the respective partial view, further improving readability and maintainability.